fix: enable saved acc and refactor
