+++ src/lib/layout/navbar/Collapse.tsx
@@ -6,7 +6,7 @@ import { AppLink } from "lib/components/AppLink";
 import { CustomIcon } from "lib/components/icon";
 import { Tooltip } from "lib/components/Tooltip";
 
-import type { NavMenuProps, SubmenuInfo } from "./type";
+import type { NavMenuProps, SubmenuInfo } from "./types";
 
 interface CollapseNavInfoProps {
   submenu: SubmenuInfo;
